requirements.txt
# Bibliotecas para extração de PDFs
camelot-py>=1.0.0
pdfplumber>=0.11.0

# Processamento de dados
pandas>=2.0.0

# Exportação para Excel
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# Interface gráfica
PySimpleGUI>=5.0.0

# Construção do executável
pyinstaller>=6.0.0
